% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.abram.temp.R
\name{bw.abram.temp}
\alias{bw.abram.temp}
\title{Computes adaptive smoothing bandwidth in the temporal case according to the inverse-square-root rule of Abramson (1982).}
\usage{
bw.abram.temp(t, h0 = NULL, nt = 128, trim = 5, at = "points")
}
\arguments{
\item{t}{A vector (a temporal point pattern) for which the bandwidths should be computed.}

\item{h0}{The global smoothing bandwidth. The default is Silverman's rule of thumb (bw.nrd0).}

\item{nt}{The number of equally spaced points at which the temporal density is to be estimated.}

\item{trim}{A trimming value to cut extreme large bandwidths.}

\item{at}{Character string specifying whether to compute bandwidths at the points (at = "points", the default) or to compute bandwidths at every bin in a bin grid (at = "bins").}
}
\description{
Computes adaptive smoothing bandwidth in the temporal case according to the inverse-square-root rule of Abramson (1982).
}
\examples{
X <- rbeta(100, 1.5, 5.5, 0.2) * 2 #Simulated temporal point pattern
bw.abram.temp(X)
}
